+++ src/action/output.rs
@@ -12,10 +12,7 @@ pub struct Output<Value = ()> {
   pub digested: usize,
 }
 
-impl<'a, Text: ?Sized, StateRef, HeapRef> Input<&'a Text, StateRef, HeapRef>
-where
-  &'a Text: Digest,
-{
+impl<Text: ?Sized + Digest, StateRef, HeapRef> Input<&Text, StateRef, HeapRef> {
   /// Validate if it is ok to digest `n` bytes.
   /// See [`Digest::validate`] for more information.
   #[inline]